fix: Group B — fractional shares end-to-end + test mock alignment
Fractional-shares source bug (test_reconcile × 2 failures):
- broker_interface: BrokerPosition.quantity, BrokerOrder.quantity, and
  BrokerOrder.filled_qty changed from int to float. Alpaca supports
  fractional shares; int coercion was silently truncating 0.30 → 0.
- alpaca_broker: replaced `int(float(...))` with `float(... or 0)` in
  all 4 position/order construction sites.
- reconcile._backfill_trade_data: passes float(qty) as planned_shares
  instead of int(qty).
- schema.registry: shadow_trades.planned_shares and .actual_shares
  changed from INTEGER to REAL so fractional values persist correctly.
  Migration applied via validate-schema --fix.

Test mock alignment (test_ingestion × 2, test_news, test_render_sync):
- test_ingestion: mock yf.download instead of hitting real network.
  Violated CLAUDE.md's "mock all external APIs in tests" rule and
  yfinance returns None for some periods, flaking CI.
- test_news: patch _load_cached and clear FINNHUB_API_KEY so the
  "no API key → None" branch is actually exercised, not masked by a
  stale on-disk cache from prior runs.
- test_render_sync: patch src.schema.postgres.create_all_tables +
  ensure_columns so their internal psycopg2.connect calls don't
  inflate the healthy-reconnection test's connect count from 1 → 3.

83 tests pass, 0 failures across all 5 affected test files.

+++ src/schema/registry.py
@@ -182,7 +182,10 @@ _register(TableDef(
         ColumnDef("stop_price", "REAL"),
         ColumnDef("target_1", "REAL"),
         ColumnDef("target_2", "REAL"),
-        ColumnDef("planned_shares", "INTEGER"),
+        # REAL (not INTEGER): Alpaca fractional shares. INTEGER would silently
+        # truncate e.g. 0.30 → 0 on reconcile, then the positive-shares guard
+        # in journal.store would reject the row.
+        ColumnDef("planned_shares", "REAL"),
         ColumnDef("planned_allocation", "REAL"),
         ColumnDef("actual_entry_price", "REAL"),
         ColumnDef("actual_entry_time", "TEXT"),
@@ -228,7 +231,8 @@ _register(TableDef(
         ColumnDef("fill_price", "REAL"),
         ColumnDef("implementation_shortfall_bps", "REAL"),
         ColumnDef("strategy_type", "TEXT", default="pullback"),
-        ColumnDef("actual_shares", "INTEGER"),
+        # REAL (not INTEGER): matches planned_shares — Alpaca fractional support.
+        ColumnDef("actual_shares", "REAL"),
         ColumnDef("exit_retry_count", "INTEGER", default="0"),
         # Strategy Decision #24: Outcome metadata for regime-conditional analysis.
         # These columns capture market context at entry/exit so we can answer
